Passing database configuration test.
Test wasn't functioning correctly before because I wasn't using specs2 to deal with Future objects correctly.

Still am not completely satisfied with the structure of DatabaseConfigSpecification. Could do with a test for the table drop that takes place in the afterAll method. But it's time to move on to actual bling functionality. The test improvement can come later.